src/main/java/com/damian/figinski/dentistapp/service/PatientService.java
package com.damian.figinski.dentistapp.service;

import com.damian.figinski.dentistapp.model.Patient;
import com.damian.figinski.dentistapp.repository.PatientRepository;
import org.springframework.stereotype.Service;

import java.util.List;

@Service
public class PatientService {

    private  final PatientRepository patientRepository;

    // Constructor Injection
    public PatientService(PatientRepository patientRepository) {
        this.patientRepository = patientRepository;
    }

    public List<Patient> findAll() {
        return patientRepository.findAll();
    }

    public Patient save(Patient patient) {
        return patientRepository.save(patient);
    }

    public Patient findById(Long id) {
        return patientRepository.findById(id).orElse(null);
    }

    public void deleteById(Long id) {
        patientRepository.deleteById(id);
    }
}
